src/main/java/com/nxp/hcedemomws/dto/InitPurchaseOrderRequest.java
package com.nxp.hcedemomws.dto;

import java.io.Serializable;

import com.fasterxml.jackson.annotation.JsonInclude;

@JsonInclude(JsonInclude.Include.NON_NULL)
public class InitPurchaseOrderRequest implements Serializable{

	/**
	 * 
	 */
	private static final long serialVersionUID = 2648397371275914181L;
	
	private String deviceId;
	
	private String googleAccountId;
	
	private String product;

	public String getDeviceId() {
		return deviceId;
	}

	public void setDeviceId(String deviceId) {
		this.deviceId = deviceId;
	}

	public String getGoogleAccountId() {
		return googleAccountId;
	}

	public void setGoogleAccountId(String googleAccountId) {
		this.googleAccountId = googleAccountId;
	}

	public String getProduct() {
		return product;
	}

	public void setProduct(String product) {
		this.product = product;
	}

}
